% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_pathway_activity.R
\name{calculate_aucell_scores}
\alias{calculate_aucell_scores}
\title{Calculate AUCell scores}
\usage{
calculate_aucell_scores(expression_matrix, gene_sets, ...)
}
\description{
Calculate AUCell scores
}
\keyword{internal}
