% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrate_levels.R
\name{calculate_consistency}
\alias{calculate_consistency}
\title{Calculate consistency between gene and pathway evidence}
\usage{
calculate_consistency(gene_evidence, pathway_evidence)
}
\description{
Calculate consistency between gene and pathway evidence
}
\keyword{internal}
