% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_statistical.R
\name{calculate_correlation}
\alias{calculate_correlation}
\title{Calculate correlation between two vectors}
\usage{
calculate_correlation(x, y, method = "pearson")
}
\arguments{
\item{x}{First vector}

\item{y}{Second vector}

\item{method}{Correlation method}
}
\value{
Correlation coefficient
}
\description{
Calculate correlation between two vectors
}
\keyword{internal}
