% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_pathway_activity.R
\name{calculate_gsva_scores}
\alias{calculate_gsva_scores}
\title{Calculate GSVA scores}
\usage{
calculate_gsva_scores(expression_matrix, gene_sets, parallel, ...)
}
\description{
Calculate GSVA scores
}
\keyword{internal}
