% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_pathway_activity.R
\name{calculate_mean_scores}
\alias{calculate_mean_scores}
\title{Calculate mean expression scores}
\usage{
calculate_mean_scores(expression_matrix, gene_sets)
}
\description{
Calculate mean expression scores
}
\keyword{internal}
