% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_pathway_activity.R
\name{calculate_median_scores}
\alias{calculate_median_scores}
\title{Calculate median expression scores}
\usage{
calculate_median_scores(expression_matrix, gene_sets)
}
\description{
Calculate median expression scores
}
\keyword{internal}
