% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_general.R
\name{calculate_overlap}
\alias{calculate_overlap}
\title{Find overlapping elements between lists}
\usage{
calculate_overlap(list1, list2, method = "intersect")
}
\arguments{
\item{list1}{First list}

\item{list2}{Second list}

\item{method}{Method for overlap: "intersect", "jaccard", "dice"}
}
\value{
List with overlap information
}
\description{
Find overlapping elements between lists
}
\keyword{internal}
