% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_general.R
\name{calculate_quantiles}
\alias{calculate_quantiles}
\title{Calculate quantiles of a numeric vector}
\usage{
calculate_quantiles(x, probs = c(0.25, 0.5, 0.75), na.rm = TRUE)
}
\arguments{
\item{x}{Numeric vector}

\item{probs}{Quantile probabilities}

\item{na.rm}{Remove NA values}
}
\value{
Named numeric vector of quantiles
}
\description{
Calculate quantiles of a numeric vector
}
\keyword{internal}
