% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_statistical.R
\name{calculate_size_factors}
\alias{calculate_size_factors}
\title{Calculate size factors using DESeq2 method}
\usage{
calculate_size_factors(counts)
}
\arguments{
\item{counts}{Count matrix}
}
\value{
Size factors
}
\description{
Calculate size factors using DESeq2 method
}
\keyword{internal}
