% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_pathway_activity.R
\name{calculate_ssgsea_scores}
\alias{calculate_ssgsea_scores}
\title{Calculate ssGSEA scores}
\usage{
calculate_ssgsea_scores(expression_matrix, gene_sets, parallel, ...)
}
\description{
Calculate ssGSEA scores
}
\keyword{internal}
