% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze_regions.R
\name{calculate_transcript_proportions}
\alias{calculate_transcript_proportions}
\title{Calculate Transcript Usage Proportions}
\usage{
calculate_transcript_proportions(counts, annotation)
}
\arguments{
\item{counts}{Transcript count matrix}

\item{annotation}{Data.frame with transcript_id and gene_id columns}
}
\value{
Matrix of transcript usage proportions
}
\description{
Calculates the proportion of each transcript relative to total gene expression
}
\examples{
proportions <- calculate_transcript_proportions(transcript_counts, annotation)

}
