% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_validation.R
\name{check_package}
\alias{check_package}
\title{Check if package is available}
\usage{
check_package(package, function_name = "This function")
}
\arguments{
\item{package}{Package name}

\item{function_name}{Function name for error message}
}
\value{
TRUE if available, stops if not
}
\description{
Check if package is available
}
\keyword{internal}
