% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_general.R
\name{convert_gene_ids}
\alias{convert_gene_ids}
\title{Convert gene IDs between different formats}
\usage{
convert_gene_ids(
  genes,
  from = "symbol",
  to = "ensembl",
  species = "human",
  remove_unmapped = TRUE
)
}
\arguments{
\item{genes}{Character vector of gene identifiers}

\item{from}{Source identifier type: "symbol", "ensembl", "entrez"}

\item{to}{Target identifier type: "symbol", "ensembl", "entrez"}

\item{species}{Character, "human" or "mouse"}

\item{remove_unmapped}{Logical, remove genes that couldn't be mapped}
}
\value{
Named character vector with converted IDs
}
\description{
Convert gene IDs between different formats
}
