% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_annotation.R
\name{create_gene_annotation}
\alias{create_gene_annotation}
\title{Create gene annotation data.frame}
\usage{
create_gene_annotation(
  genes,
  species = "human",
  id_type = "symbol",
  include_biotype = TRUE,
  include_chromosome = FALSE
)
}
\arguments{
\item{genes}{Character vector of gene identifiers}

\item{species}{Character, "human" or "mouse"}

\item{id_type}{Type of input IDs}

\item{include_biotype}{Include biotype information}

\item{include_chromosome}{Include chromosome information}
}
\value{
Data.frame with gene annotations
}
\description{
Create gene annotation data.frame
}
