% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze_families.R
\name{create_gene_mapping}
\alias{create_gene_mapping}
\title{Convert family definitions to gene mapping}
\usage{
create_gene_mapping(families)
}
\description{
Convert family definitions to gene mapping
}
\keyword{internal}
