% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_results.R
\name{create_html_report_content}
\alias{create_html_report_content}
\title{Create HTML report content}
\usage{
create_html_report_content(
  results_list,
  include_plots,
  significant_only,
  alpha,
  metadata
)
}
\description{
Create HTML report content
}
\keyword{internal}
