% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrate_levels.R
\name{create_integration_summary}
\alias{create_integration_summary}
\title{Create integration summary}
\usage{
create_integration_summary(results_list, integrated, alpha)
}
\description{
Create integration summary
}
\keyword{internal}
