% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_general.R
\name{create_named_vector}
\alias{create_named_vector}
\title{Create a named vector from two vectors}
\usage{
create_named_vector(names, values, remove_na = TRUE)
}
\arguments{
\item{names}{Character vector of names}

\item{values}{Vector of values}

\item{remove_na}{Remove NA values}
}
\value{
Named vector
}
\description{
Create a named vector from two vectors
}
\keyword{internal}
