% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_annotation.R
\name{create_pathway_matrix}
\alias{create_pathway_matrix}
\title{Create gene-pathway membership matrix}
\usage{
create_pathway_matrix(pathways, genes)
}
\arguments{
\item{pathways}{Named list of pathways}

\item{genes}{Character vector of all genes}
}
\value{
Binary matrix (genes x pathways)
}
\description{
Create gene-pathway membership matrix
}
\keyword{internal}
