% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_pathway_heatmap.R
\name{create_pheatmap}
\alias{create_pheatmap}
\title{Create heatmap using pheatmap}
\usage{
create_pheatmap(
  plot_data,
  metadata,
  annotation_cols,
  color_scheme,
  scale,
  cluster_rows,
  cluster_cols,
  show_pathway_names,
  fontsize_row,
  fontsize_col,
  ...
)
}
\description{
Create heatmap using pheatmap
}
\keyword{internal}
