% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_results.R
\name{create_summary_for_export}
\alias{create_summary_for_export}
\title{Create summary for export}
\usage{
create_summary_for_export(results_list, alpha)
}
\description{
Create summary for export
}
\keyword{internal}
