% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_results.R
\name{export_to_excel}
\alias{export_to_excel}
\title{Export to Excel format}
\usage{
export_to_excel(
  results_list,
  output_dir,
  prefix,
  separate_sheets,
  significant_only,
  alpha,
  ...
)
}
\description{
Export to Excel format
}
\keyword{internal}
