% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_results.R
\name{export_to_html}
\alias{export_to_html}
\title{Export to HTML format}
\usage{
export_to_html(
  results_list,
  output_dir,
  prefix,
  include_plots,
  significant_only,
  alpha,
  metadata,
  ...
)
}
\description{
Export to HTML format
}
\keyword{internal}
