% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_annotation.R
\name{filter_genes_by_biotype}
\alias{filter_genes_by_biotype}
\title{Filter genes by biotype}
\usage{
filter_genes_by_biotype(
  genes,
  biotypes = "protein_coding",
  species = "human",
  id_type = "symbol"
)
}
\arguments{
\item{genes}{Character vector of gene identifiers}

\item{biotypes}{Character vector of biotypes to keep}

\item{species}{Character, "human" or "mouse"}

\item{id_type}{Type of input IDs}
}
\value{
Filtered character vector of genes
}
\description{
Filter genes by biotype
}
