% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_annotation.R
\name{filter_pathways_to_genes}
\alias{filter_pathways_to_genes}
\title{Get pathway-gene mappings from annotation}
\usage{
filter_pathways_to_genes(pathways, genes)
}
\arguments{
\item{pathways}{Named list of pathways}

\item{genes}{Character vector of genes to filter to}
}
\value{
List with filtered pathways
}
\description{
Get pathway-gene mappings from annotation
}
\keyword{internal}
