% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_general.R
\name{format_logfc}
\alias{format_logfc}
\title{Format log fold changes for display}
\usage{
format_logfc(lfc, digits = 2)
}
\arguments{
\item{lfc}{Numeric vector of log fold changes}

\item{digits}{Number of decimal places}
}
\value{
Character vector of formatted LFCs
}
\description{
Format log fold changes for display
}
\keyword{internal}
