% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_annotation.R
\name{get_gene_biotypes}
\alias{get_gene_biotypes}
\title{Get gene biotype information}
\usage{
get_gene_biotypes(genes, species = "human", id_type = "symbol")
}
\arguments{
\item{genes}{Character vector of gene identifiers}

\item{species}{Character, "human" or "mouse"}

\item{id_type}{Type of input IDs: "symbol", "ensembl", "entrez"}
}
\value{
Data.frame with gene and biotype columns
}
\description{
Get gene biotype information
}
\keyword{internal}
