% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_annotation.R
\name{get_gene_chromosomes}
\alias{get_gene_chromosomes}
\title{Get chromosome information for genes}
\usage{
get_gene_chromosomes(genes, species = "human", id_type = "symbol")
}
\arguments{
\item{genes}{Character vector of gene identifiers}

\item{species}{Character, "human" or "mouse"}

\item{id_type}{Type of input IDs: "symbol", "ensembl", "entrez"}
}
\value{
Data.frame with gene and chromosome columns
}
\description{
Get chromosome information for genes
}
\keyword{internal}
