% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_annotation.R
\name{get_gene_descriptions}
\alias{get_gene_descriptions}
\title{Get gene descriptions/names}
\usage{
get_gene_descriptions(genes, species = "human", id_type = "symbol")
}
\arguments{
\item{genes}{Character vector of gene identifiers}

\item{species}{Character, "human" or "mouse"}

\item{id_type}{Type of input IDs}
}
\value{
Data.frame with gene and description columns
}
\description{
Get gene descriptions/names
}
\keyword{internal}
