% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_annotation.R
\name{get_gene_lengths}
\alias{get_gene_lengths}
\title{Get gene lengths for TPM/FPKM calculations}
\usage{
get_gene_lengths(genes, species = "human", id_type = "symbol")
}
\arguments{
\item{genes}{Character vector of gene identifiers}

\item{species}{Character, "human" or "mouse"}

\item{id_type}{Type of input IDs}
}
\value{
Named numeric vector of gene lengths
}
\description{
Get gene lengths for TPM/FPKM calculations
}
\keyword{internal}
