% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pathways.R
\name{get_kegg_latest}
\alias{get_kegg_latest}
\title{Get latest KEGG pathways using KEGG REST API}
\usage{
get_kegg_latest(
  species = "human",
  min_size = 5,
  max_size = 500,
  id_type = "symbol"
)
}
\description{
Get latest KEGG pathways using KEGG REST API
}
\keyword{internal}
