% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_general.R
\name{get_memory_usage}
\alias{get_memory_usage}
\title{Get memory usage in human readable format}
\usage{
get_memory_usage(object)
}
\arguments{
\item{object}{Object to measure}
}
\value{
Character string with memory usage
}
\description{
Get memory usage in human readable format
}
\keyword{internal}
