% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_general.R
\name{get_top_n}
\alias{get_top_n}
\title{Get top N elements from named vector}
\usage{
get_top_n(x, n = 10, decreasing = TRUE)
}
\arguments{
\item{x}{Named numeric vector}

\item{n}{Number of elements to return}

\item{decreasing}{Sort in decreasing order}
}
\value{
Named vector with top N elements
}
\description{
Get top N elements from named vector
}
\keyword{internal}
