% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_general.R
\name{has_class}
\alias{has_class}
\title{Check if object has specific class}
\usage{
has_class(x, class_name)
}
\arguments{
\item{x}{Object to check}

\item{class_name}{Name of class}
}
\value{
Logical
}
\description{
Check if object has specific class
}
\keyword{internal}
