% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrate_levels.R
\name{integrate_consensus}
\alias{integrate_consensus}
\title{Consensus integration}
\usage{
integrate_consensus(results_list, alpha, min_levels, ...)
}
\description{
Consensus integration
}
\keyword{internal}
