% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrate_levels.R
\name{integrate_evidence_scoring}
\alias{integrate_evidence_scoring}
\title{Evidence scoring integration}
\usage{
integrate_evidence_scoring(results_list, weights, alpha, ...)
}
\description{
Evidence scoring integration
}
\keyword{internal}
