% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrate_levels.R
\name{integrate_pathway_guided}
\alias{integrate_pathway_guided}
\title{Pathway-guided integration}
\usage{
integrate_pathway_guided(results_list, pathway_db, alpha, ...)
}
\description{
Pathway-guided integration
}
\keyword{internal}
