% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrate_levels.R
\name{integrate_rank_aggregation}
\alias{integrate_rank_aggregation}
\title{Rank aggregation integration}
\usage{
integrate_rank_aggregation(results_list, weights, alpha, ...)
}
\description{
Rank aggregation integration
}
\keyword{internal}
