% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_statistical.R
\name{normalize_counts}
\alias{normalize_counts}
\title{Normalize counts using size factors}
\usage{
normalize_counts(counts, size_factors = NULL)
}
\arguments{
\item{counts}{Count matrix}

\item{size_factors}{Size factors}
}
\value{
Normalized counts
}
\description{
Normalize counts using size factors
}
\keyword{internal}
