% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_multilevel.R
\name{plot_concordance}
\alias{plot_concordance}
\title{Plot concordance between levels}
\usage{
plot_concordance(results_list, alpha, lfc_threshold, color_palette, ...)
}
\description{
Plot concordance between levels
}
\keyword{internal}
