% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_pathway_heatmap.R
\name{plot_pathway_heatmap_interactive}
\alias{plot_pathway_heatmap_interactive}
\title{Interactive Pathway Heatmap}
\usage{
plot_pathway_heatmap_interactive(
  pathway_results,
  metadata = NULL,
  top_n = 20,
  scale = "row",
  color_scheme = "RdBu"
)
}
\arguments{
\item{pathway_results}{Output from analyze_pathways() or pathway scores matrix}

\item{metadata}{Sample metadata (optional)}

\item{top_n}{Number of top pathways to include (default: 20)}

\item{scale}{Scale data: "row", "column", "none" (default: "row")}

\item{color_scheme}{Color scheme (default: "RdBu")}
}
\value{
A plotly object
}
\description{
Creates an interactive heatmap using plotly for exploring pathway activity
}
\examples{
pathway_results <- analyze_pathways(...)
interactive_heatmap <- plot_pathway_heatmap_interactive(pathway_results)

}
