% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_multilevel.R
\name{plot_summary_bar}
\alias{plot_summary_bar}
\title{Plot summary bar chart}
\usage{
plot_summary_bar(results_list, alpha, lfc_threshold, color_palette, ...)
}
\description{
Plot summary bar chart
}
\keyword{internal}
