% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_multilevel.R
\name{plot_upset_multilevel}
\alias{plot_upset_multilevel}
\title{Plot UpSet diagram for multilevel overlap}
\usage{
plot_upset_multilevel(results_list, alpha, lfc_threshold, ...)
}
\description{
Plot UpSet diagram for multilevel overlap
}
\keyword{internal}
