% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_multilevel.R
\name{plot_volcano_grid}
\alias{plot_volcano_grid}
\title{Plot grid of volcano plots}
\usage{
plot_volcano_grid(results_list, alpha, lfc_threshold, ...)
}
\description{
Plot grid of volcano plots
}
\keyword{internal}
