% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze_genes.R
\name{print.functionalDE_result}
\alias{print.functionalDE_result}
\title{Print method for functionalDE_result}
\usage{
\method{print}{functionalDE_result}(x, ...)
}
\arguments{
\item{x}{A functionalDE_result object}

\item{...}{Additional arguments (ignored)}
}
\description{
Print method for functionalDE_result
}
