% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrate_levels.R
\name{print.funDE_integration}
\alias{print.funDE_integration}
\title{Print method for funDE_integration}
\usage{
\method{print}{funDE_integration}(x, ...)
}
\arguments{
\item{x}{A funDE_integration object}

\item{...}{Additional arguments (ignored)}
}
\description{
Print method for funDE_integration
}
