% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_pseudobulk.R
\name{print.funDE_pseudobulk}
\alias{print.funDE_pseudobulk}
\title{Print method for funDE_pseudobulk}
\usage{
\method{print}{funDE_pseudobulk}(x, ...)
}
\arguments{
\item{x}{A funDE_pseudobulk object}

\item{...}{Additional arguments (ignored)}
}
\description{
Print method for funDE_pseudobulk
}
