% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_pseudobulk.R
\name{qc_pseudobulk}
\alias{qc_pseudobulk}
\title{Quality Control for Pseudobulk Data}
\usage{
qc_pseudobulk(pseudobulk_result, plot = TRUE)
}
\arguments{
\item{pseudobulk_result}{Output from aggregate_pseudobulk()}

\item{plot}{Logical, create QC plots (default: TRUE)}
}
\value{
List with QC metrics and plots
}
\description{
Performs quality control checks on pseudobulk data to identify potential issues
}
\examples{
pseudobulk <- aggregate_pseudobulk(...)
qc_results <- qc_pseudobulk(pseudobulk)

}
