% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_statistical.R
\name{robust_stats}
\alias{robust_stats}
\title{Calculate robust statistics}
\usage{
robust_stats(x)
}
\arguments{
\item{x}{Numeric vector}
}
\value{
List of robust statistics
}
\description{
Calculate robust statistics
}
\keyword{internal}
