% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze_genes.R
\name{run_deseq2}
\alias{run_deseq2}
\title{Run DESeq2 analysis}
\usage{
run_deseq2(
  counts,
  metadata,
  design,
  contrast,
  alpha,
  lfc_threshold,
  parallel,
  ...
)
}
\description{
Run DESeq2 analysis
}
\keyword{internal}
