% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_general.R
\name{safe_extract}
\alias{safe_extract}
\title{Safely extract elements from list with defaults}
\usage{
safe_extract(x, name, default = NULL)
}
\arguments{
\item{x}{List}

\item{name}{Element name}

\item{default}{Default value if element doesn't exist}
}
\value{
Element value or default
}
\description{
Safely extract elements from list with defaults
}
\keyword{internal}
