% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_general.R
\name{safe_file_path}
\alias{safe_file_path}
\title{Create file paths safely}
\usage{
safe_file_path(..., create_dirs = FALSE)
}
\arguments{
\item{...}{Path components}

\item{create_dirs}{Create directories if they don't exist}
}
\value{
Character path
}
\description{
Create file paths safely
}
\keyword{internal}
