% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_general.R
\name{smart_round}
\alias{smart_round}
\title{Round numbers intelligently}
\usage{
smart_round(x, digits = 3)
}
\arguments{
\item{x}{Numeric vector}

\item{digits}{Number of digits}
}
\value{
Rounded numeric vector
}
\description{
Round numbers intelligently
}
\keyword{internal}
