% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze_genes.R
\name{summary.functionalDE_result}
\alias{summary.functionalDE_result}
\title{Summary method for functionalDE_result}
\usage{
\method{summary}{functionalDE_result}(object, ...)
}
\arguments{
\item{object}{A functionalDE_result object}

\item{...}{Additional arguments (ignored)}
}
\description{
Summary method for functionalDE_result
}
