% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrate_levels.R
\name{summary.funDE_integration}
\alias{summary.funDE_integration}
\title{Summary method for funDE_integration}
\usage{
\method{summary}{funDE_integration}(object, ...)
}
\arguments{
\item{object}{A funDE_integration object}

\item{...}{Additional arguments (ignored)}
}
\description{
Summary method for funDE_integration
}
