% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_pseudobulk.R
\name{summary.funDE_pseudobulk}
\alias{summary.funDE_pseudobulk}
\title{Summary method for funDE_pseudobulk}
\usage{
\method{summary}{funDE_pseudobulk}(object, ...)
}
\arguments{
\item{object}{A funDE_pseudobulk object}

\item{...}{Additional arguments (ignored)}
}
\description{
Summary method for funDE_pseudobulk
}
