% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze_pathways.R
\name{test_pathways_limma}
\alias{test_pathways_limma}
\title{Test pathway differential activity using limma}
\usage{
test_pathways_limma(pathway_scores, metadata, design, contrast)
}
\description{
Test pathway differential activity using limma
}
\keyword{internal}
