% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze_pathways.R
\name{test_pathways_ttest}
\alias{test_pathways_ttest}
\title{Test pathway differential activity using t-test}
\usage{
test_pathways_ttest(pathway_scores, metadata, contrast)
}
\description{
Test pathway differential activity using t-test
}
\keyword{internal}
