% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze_pathways.R
\name{test_pathways_wilcox}
\alias{test_pathways_wilcox}
\title{Test pathway differential activity using Wilcoxon test}
\usage{
test_pathways_wilcox(pathway_scores, metadata, contrast)
}
\description{
Test pathway differential activity using Wilcoxon test
}
\keyword{internal}
