% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_statistical.R
\name{transform_counts}
\alias{transform_counts}
\title{Perform variance stabilizing transformation}
\usage{
transform_counts(counts, method = "log2")
}
\arguments{
\item{counts}{Count matrix}

\item{method}{Method: "log2", "rlog", "vst"}
}
\value{
Transformed matrix
}
\description{
Perform variance stabilizing transformation
}
\keyword{internal}
