% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_validation.R
\name{validate_contrast}
\alias{validate_contrast}
\title{Validate contrast specification}
\usage{
validate_contrast(contrast, metadata)
}
\arguments{
\item{contrast}{Contrast specification}

\item{metadata}{Sample metadata}
}
\value{
Validated contrast
}
\description{
Validate contrast specification
}
\keyword{internal}
