% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_validation.R
\name{validate_counts}
\alias{validate_counts}
\title{Validate count matrix}
\usage{
validate_counts(counts, allow_missing = FALSE)
}
\arguments{
\item{counts}{Count matrix}

\item{allow_missing}{Allow missing values}
}
\value{
Validated count matrix
}
\description{
Validate count matrix
}
\keyword{internal}
