% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_validation.R
\name{validate_design}
\alias{validate_design}
\title{Validate design formula}
\usage{
validate_design(design, metadata)
}
\arguments{
\item{design}{Design formula}

\item{metadata}{Sample metadata}
}
\value{
Validated design
}
\description{
Validate design formula
}
\keyword{internal}
