% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_validation.R
\name{validate_gene_sets}
\alias{validate_gene_sets}
\title{Validate gene sets}
\usage{
validate_gene_sets(gene_sets, min_size = 1, max_size = Inf)
}
\arguments{
\item{gene_sets}{Named list of gene sets}

\item{min_size}{Minimum set size}

\item{max_size}{Maximum set size}
}
\value{
Validated gene sets
}
\description{
Validate gene sets
}
\keyword{internal}
