% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_validation.R
\name{validate_metadata}
\alias{validate_metadata}
\title{Validate metadata}
\usage{
validate_metadata(metadata, counts)
}
\arguments{
\item{metadata}{Sample metadata data.frame}

\item{counts}{Count matrix for sample matching}
}
\value{
Validated metadata
}
\description{
Validate metadata
}
\keyword{internal}
